+++ kits/python/agent.py
@@ -1,5 +1,5 @@
 from lux.kit import obs_to_game_state, GameState, EnvConfig
-from lux.utils import direction_to
+from lux.utils import direction_to, my_turn_to_place_factory
 import numpy as np
 import sys
 class Agent():
@@ -23,9 +23,11 @@ class Agent():
 
             # how many factories you have left to place
             factories_to_place = game_state.teams[self.player].factories_to_place
-            if factories_to_place > 0:
-                # we will spawn our factory in a random location with 100 metal and water
-                potential_spawns = game_state.board.spawns[self.player]
+            # whether it is your turn to place a factory
+            my_turn_to_place = my_turn_to_place_factory(game_state.teams[self.player].place_first, step)
+            if factories_to_place > 0 and my_turn_to_place:
+                # we will spawn our factory in a random location with 100 metal and water if it is our turn to place
+                potential_spawns = np.array(list(zip(*np.where(obs["board"]["valid_spawns_mask"] == 1))))
                 spawn_loc = potential_spawns[np.random.randint(0, len(potential_spawns))]
                 return dict(spawn=spawn_loc, metal=100, water=100)
             return dict()
@@ -34,6 +36,7 @@ class Agent():
         actions = dict()
         game_state = obs_to_game_state(step, self.env_cfg, obs)
         factories = game_state.factories[self.player]
+        game_state.teams[self.player].place_first
         factory_tiles, factory_units = [], []
         for unit_id, factory in factories.items():
             if factory.power >= self.env_cfg.ROBOTS["HEAVY"].POWER_COST and \