+++ kits/python/lux/team.py
@@ -21,7 +21,7 @@ class FactionTypes(Enum):
     FirstMars = FactionInfo(color="red", faction_id=4)
 
 class Team:
-    def __init__(self, team_id: int, agent: str, faction: FactionTypes = None, water=0, metal=0, factories_to_place=0, factory_strains=[]) -> None:
+    def __init__(self, team_id: int, agent: str, faction: FactionTypes = None, water=0, metal=0, factories_to_place=0, factory_strains=[], place_first=False) -> None:
         self.faction = faction
         self.team_id = team_id
         # the key used to differentiate ownership of things in state
@@ -31,6 +31,9 @@ class Team:
         self.metal = metal
         self.factories_to_place = factories_to_place
         self.factory_strains = factory_strains
+        # whether this team gets to place factories down first or not. The bid winner has this set to True. 
+        # If tied, player_0's team has this True
+        self.place_first = place_first
     def state_dict(self):
         return dict(
             team_id=self.team_id,
@@ -38,7 +41,9 @@ class Team:
             # note for optimization, water,metal, factories_to_place doesn't change after the early game.
             water=self.init_water,
             metal=self.init_metal,
-            factories_to_place=self.factories_to_place
+            factories_to_place=self.factories_to_place,
+            factory_strains=self.factory_strains,
+            place_first=self.place_first,
         )
     def __str__(self) -> str:
         out = f"[Player {self.team_id}]"