+++ ChangeLog.md
@@ -2,6 +2,17 @@
 
 ### v1.1.0
 
+Environment:
+- Assymmetric Maps
+- Bids are now for who goes first, not who gets an extra factory. If bid is tied, player 0 goes first. This early phase is kept parallel for simplicity. Instead, when it is not your turn to place a factory you must skip your turn. Subsequent turns after the starting phase are still in parallel.
+- renamed `board.spawns` -> `board.valid_spawns_mask`. Stores the same information (locations on map that you can spawn a factory on). This changes over time based on where new factories are placed.
+- Factories can spawn anywhere, but must not overlap any resource or factory.
+- Switch to x,y indexing
+
+Python Kits:
+- Updated to accept some new observation entires (`teams[player].place_first`, `board.valid_spawns_mask`), and removed old ones (`board.spawns`)
+- Added utility function to determine if it is your team's turn to place a factory or not
+
 Bug Fixes:
 - Fix bug where move center costs power
 - Fix extra line in stderr logging